% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT.R
\name{IRTparams}
\alias{IRTparams}
\title{Gets/sets the parameters of the model}
\usage{
IRTparams(model)
}
\arguments{
\item{model}{An object of class IRTModel}
}
\value{
a list of parameters
}
\description{
Gets/sets the parameters of the model
}
