% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT.R
\name{logit}
\alias{logit}
\alias{ilogit}
\alias{lilogit}
\title{Logistic and inverse logistic function}
\usage{
logit(p)

ilogit(x)

lilogit(x)
}
\arguments{
\item{p}{-- The probability value to transform to theta space.}

\item{x}{-- The theta value to transform into probability space.}
}
\value{
For \code{logit} the scale value for \code{ilogit}
(\code{lilogit}) the (log) of the probability.
}
\description{
The function \code{logit} is the logistic function, $\log p/(1-p)$.
The function \code{ilogit} is the inverse logistic function,
$1/(1+e^{-x})$.  The function \code{lilogit} is the negative
log of the inverse logit
}
\section{Functions}{
\itemize{
\item \code{ilogit()}: Inverse logit

\item \code{lilogit()}: log inverse logit

}}
